+++ npe2/cli.py
@@ -475,7 +475,7 @@ def cache(
 
 @app.command()
 def compile(src_dir: str):
-    """Compile napari_plugin_engine plugins"""
+    """Compile napari_plugin_engine plugins."""
     from .implements import compile
 
     _pprint_formatted(compile(src_dir).yaml(), Format.yaml)
