+++ tests/conftest.py
@@ -1,3 +1,4 @@
+import shutil
 import sys
 from importlib import abc, metadata
 from pathlib import Path
@@ -8,6 +9,8 @@ import pytest
 from npe2 import PluginManager, PluginManifest
 from npe2.manifest import _npe1_adapter
 
+FIXTURES = Path(__file__).parent / "fixtures"
+
 
 @pytest.fixture
 def sample_path():
@@ -17,6 +20,12 @@ def sample_path():
 @pytest.fixture
 def sample_manifest(sample_path):
     return PluginManifest.from_file(sample_path / "my_plugin" / "napari.yaml")
+
+
+@pytest.fixture
+def compiled_plugin_dir(tmp_path):
+    shutil.copytree(FIXTURES / "my-compiled-plugin", tmp_path, dirs_exist_ok=True)
+    return tmp_path
 
 
 @pytest.fixture
